+++ NEWS.md
@@ -1,4 +1,5 @@
 # glitr 0.1.0
+* Added a new masked dataset for training/demoing, `hfr_mmd`
 * New parameter in`si_sytle()` to allow for all black text
   * `FO = TRUE` sets all text to be black throughout